musiccast: split into player.py, provider.py
